Move MethodIdentifier
