+++ transaction/src/model/mod.rs
@@ -7,7 +7,6 @@ mod notarized_transaction;
 mod preview_transaction;
 mod system_transaction;
 mod test_transaction;
-mod validated_transaction;
 
 pub use self::notarized_transaction::*;
 pub use auth_module::*;
@@ -18,4 +17,3 @@ pub use manifest::*;
 pub use preview_transaction::*;
 pub use system_transaction::*;
 pub use test_transaction::*;
-pub use validated_transaction::*;
